# api/openapi.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "AT Protocol Music API",
    "version": "1.0.0",
    "description": "Frozen MVP API contract from docs/planning/API.md and MVP.md. A route in this specification is not a claim that the current executable implements it. Schema and transport fixtures use reserved identities and a fixture-only namespace; production publication requires an owner-controlled prefix. All application JSON objects are closed; absent optional fields are omitted, not null/empty. No upstream access/refresh tokens or private signing keys are returned."
  },
  "servers": [
    {
      "url": "/",
      "description": "Configured application origin; never derive OAuth callback from Host headers."
    }
  ],
  "paths": {
...
    "/api/v1/resolve": {
      "get": {
        "operationId": "resolve",
        "summary": "Resolve",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedIdentity"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/resolve"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The resource is absent or inactive.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-404"
                  }
                }
              }
            }
          },
          "422": {
            "description": "A request field is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-422"
                  }
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-429"
                  }
                }
              }
            }
          },
          "502": {
            "description": "The upstream request permanently failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-502"
                  }
                }
              }
            }
          }
        },
        "x-auth-mode": "public",
        "description": "Verified handle-to-DID round trip; unresolved 404, mismatch 422. No unbounded implicit discovery for unknown collection owners.",
        "parameters": [
          {
            "$ref": "#/components/parameters/HandleQuery"
          }
        ]
      }
    },
...
    "/api/v1/scrobbles/{id}": {
      "get": {
        "operationId": "scrobble_get",
        "summary": "Scrobble get",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrobbleResult"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/scrobble"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The resource is absent or inactive.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-404"
                  }
                }
              }
            }
          },
          "422": {
            "description": "A request field is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-422"
                  }
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-429"
                  }
                }
              }
            }
          }
        },
        "x-auth-mode": "public",
        "description": "Inactive, tombstoned, unconfirmed and absent records return 404.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScrobbleIdPath"
          }
        ]
      },
      "delete": {
        "operationId": "scrobble_delete",
        "summary": "Scrobble delete",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/accepted"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No response body; action confirmed or already absent for authenticated owner."
          },
          "401": {
            "description": "A valid session or OAuth authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-401"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ownership, Origin or CSRF check failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-403"
                  }
                }
              }
            }
          },
          "422": {
            "description": "A request field is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-422"
                  }
                }
              }
            }
          },
          "429": {
            "description": "The request rate limit was exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-429"
                  }
                }
              }
            }
          },
          "502": {
            "description": "Error, including a repeated terminal failed removal. deletion_failed includes the original operation Location; no new operation is enqueued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "fixture": {
                    "$ref": "#/components/examples/error-502"
                  },
                  "terminalDeletionFailed": {
                    "$ref": "#/components/examples/deletion-failed"
                  }
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Original owner-visible terminal operation; required when error.code is deletion_failed.",
                "schema": {
                  "type": "string",
                  "pattern": "^/api/v1/operations/[^/?#\\s]+$"
                },
                "x-required-for-error-codes": [
                  "deletion_failed"
                ]
              }
            }
          }
        },
        "x-auth-mode": "mutation",
        "description": "Cross-owner requests are 403. Authenticated URI owner can repeat a confirmed deletion with 204. Pending deletion hides the public record immediately; unresolved remote deletion remains 202.",
        "x-csrf-required": true,
        "parameters": [
          {
            "$ref": "#/components/parameters/ScrobbleIdPath"
          },
          {
            "$ref": "#/components/parameters/Origin"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Timestamp": {
        "type": "string",
        "format": "at-utc-datetime",
        "description": "UTC RFC3339 within the AT Lexicon datetime subset: uppercase T/Z (or +00:00), valid calendar/time with seconds 00–59, arbitrary fractional precision, at or after 1970-01-01T00:00:00Z."
      },
      "Did": {
        "type": "string",
        "format": "at-did",
        "maxLength": 2048,
        "description": "Syntactic AT Protocol DID; handles and URLs are not DIDs. This shape check does not verify a method-specific identifier or resolve identity."
      },
      "AtUri": {
        "type": "string",
        "format": "at-record-uri",
        "description": "Full record AT URI with DID authority, collection NSID and record key."
      },
      "Cid": {
        "type": "string",
        "format": "at-record-cid",
        "description": "CIDv1 DAG-CBOR SHA-256 record content identifier. Schema fixtures are not verified repository evidence."
      },
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
      },
      "MusicText": {
        "type": "string",
        "format": "music-text",
        "description": "Require at most 256 raw Unicode scalar values and 1,024 raw UTF-8 bytes, then require nonempty text after trimming surrounding Unicode White_Space. Local records store trimmed text; validated remote records retain original display spelling. Empty provided album is invalid.",
        "maxLength": 256,
        "x-maxUtf8Bytes": 1024
      },
      "Handle": {
        "type": "string",
        "format": "at-handle",
        "maxLength": 253
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "$ref": "#/components/schemas/Uuid"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ResolvedIdentity": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "did",
          "handle",
          "verified"
        ],
        "properties": {
          "did": {
            "$ref": "#/components/schemas/Did"
          },
          "handle": {
            "$ref": "#/components/schemas/Handle"
          },
          "verified": {
            "const": true
          }
        }
      },
...
      "Scrobble": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "uri",
          "cid",
          "did",
          "revision",
          "artist",
          "track",
          "listenedAt",
          "createdAt",
          "indexedAt"
        ],
        "properties": {
          "uri": {
            "$ref": "#/components/schemas/AtUri"
          },
          "cid": {
            "$ref": "#/components/schemas/Cid"
          },
          "did": {
            "$ref": "#/components/schemas/Did"
          },
          "revision": {
            "type": "string"
          },
          "artist": {
            "$ref": "#/components/schemas/MusicText"
          },
          "track": {
            "$ref": "#/components/schemas/MusicText"
          },
          "album": {
            "$ref": "#/components/schemas/MusicText"
          },
          "listenedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "durationSeconds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 86400,
            "x-jsonIntegerRepresentation": true
          },
          "recordingMbid": {
            "$ref": "#/components/schemas/Uuid"
          },
          "indexedAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "Accepted": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "operationId",
          "state"
        ],
        "properties": {
          "operationId": {
            "type": "string"
          },
          "state": {
            "const": "pending"
          }
        }
      },
...
      "ScrobbleResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "scrobble"
        ],
        "properties": {
          "scrobble": {
            "$ref": "#/components/schemas/Scrobble"
          }
        }
      },
...
    },
    "parameters": {
...
      "ScrobbleIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/AtUri"
        },
        "description": "URL-encoded full AT URI; not a bare record key or operation ID."
      },
...
      "Origin": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Must equal the configured allowed application origin. Missing or invalid Origin is 403 before mutation or any outbound write."
      },
...
      "HandleQuery": {
        "name": "handle",
        "in": "query",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Handle"
        }
      },
...
    },
    "examples": {
...
      "resolve": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/resolve.json"
      },
      "scrobble": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/scrobble.json"
      },
...
      "accepted": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/accepted.json"
      },
...
      "error-401": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-401.json"
      },
      "error-403": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-403.json"
      },
      "error-404": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-404.json"
      },
...
      "error-422": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-422.json"
      },
      "error-429": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-429.json"
      },
      "error-502": {
        "summary": "Reserved offline schema fixture; not public publication evidence.",
        "externalValue": "examples/error-502.json"
      },
...
      "deletion-failed": {
        "summary": "Repeat of a terminal failed deletion",
        "externalValue": "./examples/error-deletion-failed.json"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "atmusic_session",
        "description": "Secure HttpOnly SameSite=Lax, seven-day absolute lifetime; invalid at exactly seven days. Missing/malformed/expired session is 401."
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required on every state-changing cookie-authenticated request along with allowed Origin. Retrieve through authenticated GET auth/session. Invalid/missing state is 403 before mutation/outbound write."
      }
    }
  },
...
}